agno-hub/requirements.txt
agno>=1.4.11
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
openai>=1.59.0
httpx>=0.28.0
pydantic>=2.10.0

